+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 4.1.0 - 2024-04-03
+### Changed
+- Rename `SettingsValue` to correct `SettingValue`
+- Add `getKeyAndValue` missing exception
+- Improve prerequisite flag evaluation type checks
+- Typo fixes
+
 ## 4.0.1 - 2024-02-13
 ### Fixed
 - `WEB` platform support.